from local

+++ fairseq/criterions/ctc.py
@@ -113,8 +113,8 @@ class CtcCriterion(FairseqCriterion):
         ).contiguous()  # (T, B, C) from the encoder
 
         if 1:
-            lprobs2 = lprobs[lprobs.size()[0]/2:]
-            lprobs = lprobs[:lprobs.size()[0]/2]
+            lprobs2 = lprobs[int(lprobs.size()[0]/2):]
+            lprobs = lprobs[:int(lprobs.size()[0]/2)]
             print(lprobs.size())
             print(lprobs2.size())
         if "src_lengths" in sample["net_input"]:
